Move bulk of project service logic to node core
