+++ packages/node-core/package.json
@@ -36,6 +36,7 @@
     "prom-client": "^14.0.1",
     "sequelize": "6.28.0",
     "source-map": "^0.7.4",
+    "tar": "^6.1.11",
     "vm2": "^3.9.9",
     "yargs": "^16.2.0"
   },
