+++ packages/node-core/src/utils/project.ts
@@ -1,12 +1,55 @@
 // Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import fs from 'fs';
+import os from 'os';
+import path from 'path';
+import {GithubReader, IPFSReader, LocalReader, Reader} from '@subql/common';
+import {getAllEntitiesRelations} from '@subql/utils';
 import {isNumber, range, uniq, without, flatten} from 'lodash';
 import {QueryTypes, Sequelize} from 'sequelize';
+import tar from 'tar';
 import {NodeConfig} from '../configure/NodeConfig';
+import {ISubqueryProject, StoreService} from '../indexer';
 import {getLogger} from '../logger';
 
 const logger = getLogger('Project-Utils');
+
+export async function prepareProjectDir(projectPath: string): Promise<string> {
+  const stats = fs.statSync(projectPath);
+  if (stats.isFile()) {
+    const sep = path.sep;
+    const tmpDir = os.tmpdir();
+    const tempPath = fs.mkdtempSync(`${tmpDir}${sep}`);
+    // Will promote errors if incorrect format/extension
+    await tar.x({file: projectPath, cwd: tempPath});
+    return tempPath.concat('/package');
+  } else if (stats.isDirectory()) {
+    return projectPath;
+  }
+  throw new Error(`Project path: ${projectPath} doesn't exist`);
+}
+
+// We cache this to avoid repeated reads from fs
+const projectEntryCache: Record<string, string> = {};
+
+export function getProjectEntry(root: string): string {
+  const pkgPath = path.join(root, 'package.json');
+  try {
+    if (!projectEntryCache[pkgPath]) {
+      const content = fs.readFileSync(pkgPath).toString();
+      const pkg = JSON.parse(content);
+      if (!pkg.main) {
+        return './dist';
+      }
+      projectEntryCache[pkgPath] = pkg.main.startsWith('./') ? pkg.main : `./${pkg.main}`;
+    }
+
+    return projectEntryCache[pkgPath];
+  } catch (err) {
+    throw new Error(`can not find package.json within directory ${root}`);
+  }
+}
 
 export async function getExistingProjectSchema(
   nodeConfig: NodeConfig,
@@ -58,4 +101,80 @@ export async function getEnumDeprecated(sequelize: Sequelize, enumTypeNameDeprec
     {replacements: [enumTypeNameDeprecated]}
   );
   return resultsDeprecated;
+}
+
+export async function updateDataSourcesEntry(
+  reader: Reader,
+  file: string,
+  root: string,
+  script: string
+): Promise<string> {
+  if (reader instanceof LocalReader) return file;
+  else if (reader instanceof IPFSReader || reader instanceof GithubReader) {
+    const outputPath = `${path.resolve(root, file.replace('ipfs://', ''))}.js`;
+    await fs.promises.writeFile(outputPath, script);
+    return outputPath;
+  }
+  throw new Error('Un-known reader type');
+}
+
+export async function updateProcessor(reader: Reader, root: string, file: string): Promise<string> {
+  if (reader instanceof LocalReader) {
+    return path.resolve(root, file);
+  } else {
+    const res = await reader.getFile(file);
+    if (!res) {
+      throw new Error(`Unable to read file ${file}`);
+    }
+    const outputPath = `${path.resolve(root, file.replace('ipfs://', ''))}.js`;
+    await fs.promises.writeFile(outputPath, res);
+    return outputPath;
+  }
+}
+
+export async function loadDataSourceScript(reader: Reader, file?: string): Promise<string> {
+  let entry = file;
+  //For RuntimeDataSourceV0_0_1
+  if (!entry) {
+    const pkg = await reader.getPkg();
+    if (pkg === undefined) throw new Error('Project package.json is not found');
+    if (pkg.main) {
+      entry = pkg.main.startsWith('./') ? pkg.main : `./${pkg.main}`;
+    } else {
+      entry = './dist';
+    }
+  }
+  //Else get file
+  const entryScript = await reader.getFile(entry);
+  if (entryScript === undefined) {
+    throw new Error(`Entry file ${entry} for datasource not exist`);
+  }
+  return entryScript;
+}
+
+async function makeTempDir(): Promise<string> {
+  const sep = path.sep;
+  const tmpDir = os.tmpdir();
+  return fs.promises.mkdtemp(`${tmpDir}${sep}`);
+}
+
+export async function getProjectRoot(reader: Reader): Promise<string> {
+  if (reader instanceof LocalReader) return reader.root;
+  if (reader instanceof IPFSReader || reader instanceof GithubReader) {
+    return makeTempDir();
+  }
+  throw new Error('Un-known reader type');
+}
+
+export async function initDbSchema(
+  project: ISubqueryProject,
+  schema: string,
+  storeService: StoreService
+): Promise<void> {
+  const modelsRelation = getAllEntitiesRelations(project.schema);
+  await storeService.init(modelsRelation, schema);
+}
+
+export async function initHotSchemaReload(schema: string, storeService: StoreService): Promise<void> {
+  await storeService.initHotSchemaReloadQueries(schema);
 }
