+++ packages/node/package.json
@@ -47,7 +47,6 @@
     "rimraf": "^3.0.2",
     "rxjs": "^7.5.2",
     "sequelize": "6.28.0",
-    "tar": "^6.1.11",
     "vm2": "^3.9.9",
     "yargs": "^16.2.0"
   },
