+++ packages/node/src/subcommands/reindex.service.ts
@@ -9,12 +9,12 @@ import {
   StoreService,
   getExistingProjectSchema,
   CacheMetadataModel,
+  initDbSchema,
 } from '@subql/node-core';
 import { Sequelize } from 'sequelize';
 import { SubqlProjectDs, SubqueryProject } from '../configure/SubqueryProject';
 import { DynamicDsService } from '../indexer/dynamic-ds.service';
 import { UnfinalizedBlocksService } from '../indexer/unfinalizedBlocks.service';
-import { initDbSchema } from '../utils/project';
 import { reindex } from '../utils/reindex';
 
 import { ForceCleanService } from './forceClean.service';
